+++ FinalPJTimeManage/app/src/main/java/entity/U234Work.java
@@ -14,7 +14,7 @@ public class U234Work extends Work{
     public U234Work() {
     }
 
-    public U234Work(String name, double time, String note, int type, boolean t2, boolean t3, boolean t4, boolean t5, boolean t6, boolean t7, boolean CN) {
+    public U234Work(String name, long time, String note, int type, boolean t2, boolean t3, boolean t4, boolean t5, boolean t6, boolean t7, boolean CN) {
         super(name, time, note, type);
         T2 = t2;
         T3 = t3;
